requirements.txt
azure-functions
requests
python-dotenv
azure-identity
azure-keyvault-secrets